requirements.txt
fastapi
uvicorn
pandas
openai
plotly
requests
numpy
mysql-connector-python
python-dotenv
pydantic
python-multipart
jinja2
httpx
PyPDF2
sentence-transformers
faiss-cpu
streamlit
scipy